Move cookie and token logic into own files
