+++ token.go
@@ -1,1 +1,24 @@
 package password
+
+import (
+	"time"
+
+	jwt "github.com/dgrijalva/jwt-go"
+)
+
+// GenToken generates a new JSON web token.
+func GenToken(id string) (string, error) {
+	jwt := jwt.New(jwt.SigningMethodHS256)
+	expTime := time.Now().Add(time.Hour * 72).Unix()
+
+	jwt.Claims["sub"] = id
+	jwt.Claims["exp"] = expTime
+	jwt.Claims["iat"] = time.Now().Unix()
+
+	tokStr, err := jwt.SignedString(signingKey)
+	if err != nil {
+		return "", err // failed to sign token
+	}
+
+	return tokStr, nil
+}
